Azureunction/requirements.txt
aiohttp==3.7.4.post0
appnope==0.1.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
avro==1.10.2
azure-core==1.16.0
azure-cosmos==4.2.0
azure-eventhub==5.6.0
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-functions==1.7.2
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
config==0.5.0.post0
cryptography==2.8
databricks-cli==0.14.3
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
Faker==8.10.3
fastavro==1.4.2
idna==2.10
ipykernel==6.0.1
ipynb-py-convert==0.4.6
ipython==7.25.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
jq==1.1.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msrest==0.6.21
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tabulate==0.8.9
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
uamqp==1.4.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
